+++ diags/workflow.py
@@ -1,3 +1,4 @@
+from __init__ import get_conf, adjust_image
 from icons import (
     AppNode, Application, Manifest, Resolver,
     Processor, ProcessorCaller, Project, ProjectPart,
@@ -9,13 +10,10 @@ if __name__ == "__main__":
 
     from diagrams import Cluster, Diagram
 
-    # Use the script filename without Python extension
-    STATIC_DIR = Path("var/diags/")
-    filename = STATIC_DIR / Path(__file__).stem
+    conf = get_conf(Path(__file__).stem)
+    output_filepath = conf.pop("output_filepath")
 
-    # Output diagram as a PNG file with direction 'Top to Bottom'.
-    with Diagram("", filename=filename, show=False, direction="TB"):
-
+    with Diagram("", **conf):
         with Cluster("Application repository"):
             app_1 = Application("Application 1")
             app_3 = Application("Application 3")
@@ -85,4 +83,5 @@ if __name__ == "__main__":
         part_b >> project
         part_a >> project
 
-    print("Built diagram at: {}".format(filename.resolve()))
+    adjust_image(output_filepath.resolve())
+    print("Built diagram at: {}".format(output_filepath.resolve()))